gcp-event-arch-repo/infra/function/requirements.txt
google-cloud-bigquery>=3.14.0,<4.0.0

